+++ src/index.ts
@@ -64,6 +64,7 @@ export { default as Fieldset } from "./components/Fieldset/Fieldset";
 // The code associated with the FilterBar component will not be exported until
 // the refactor of the v2 version of the FilterBar component has been completed.
 // export { default as FilterBar } from "./components/FilterBar/FilterBar";
+export { default as FilterBarPopup } from "./components/FilterBarPopup/FilterBarPopup";
 export { default as Form, FormField, FormRow } from "./components/Form/Form";
 export { default as Heading } from "./components/Heading/Heading";
 export type { HeadingSizes, HeadingLevels } from "./components/Heading/Heading";